Move the code gen grammer validation
the `sysl validate` command was not really its own sub-command
and should have been attached to the `sysl codegen` command instead.
Added a --validate-only flag to the codegen command which can be used
instead to just run the grammar/transform validation
