+++ sysl2/sysl/cmd_runner.go
@@ -42,7 +42,6 @@ func (r *cmdRunner) Configure(app *kingpin.Application) error {
 		&datamodelCmd{},
 		&codegenCmd{},
 		&sequenceDiagramCmd{},
-		&validateCmd{},
 		&importSwaggerCmd{},
 		&infoCmd{},
 	}
